Finalize float arithmetic fix for clock view build
Agent-Logs-Url: https://github.com/alakmar344/Not-a-thing-launcher/sessions/6854c188-a7a8-4afb-bdcc-0542790df8dd

+++ app/src/main/java/com/nothinglauncher/DotMatrixClockView.kt
@@ -173,7 +173,7 @@ class DotMatrixClockView @JvmOverloads constructor(
         val dotSpacing = dotSize * 1.4f
         val totalWidth = totalCols * dotSpacing
         var startX = (width.toFloat() - totalWidth) / 2f
-        val startY = (height - 7 * dotSpacing) / 2f
+        val startY = (height.toFloat() - 7 * dotSpacing) / 2f
 
         for (c in chars) {
             val pattern = digitPatterns[c] ?: continue
